Add `tsp` to find best path in any system
