+++ utils/systems/tsp/TSP_NA7.json
@@ -1,0 +1,7 @@
+{
+    "NA7": [
+        "NA7-GH",
+        "NA7-TH",
+        "NA7-W-ZY1"
+    ]
+}
