+++ utils/systems/tsp/TSP_OE.json
@@ -1,0 +1,14 @@
+{
+    "OE": [
+        "OE-PM",
+        "OE-PM-TR",
+        "OE-BO",
+        "OE-KO",
+        "OE-UC",
+        "OE-UC-OB",
+        "OE-UC-AD",
+        "OE-W-XV",
+        "OE-NY",
+        "OE-CR"
+    ]
+}
