+++ utils/systems/tsp/TSP_XV.json
@@ -1,0 +1,15 @@
+{
+    "XV": [
+        "XV-BN",
+        "XV-XA",
+        "XV-ST",
+        "XV-ST-BG",
+        "XV-OS",
+        "XV-CB-NM",
+        "XV-CB",
+        "XV-CB-IT",
+        "XV-W-ZY1",
+        "XV-W-OE",
+        "XV-SN"
+    ]
+}
